args-free build
