+++ bin/jspm
@@ -27,11 +27,11 @@ if (!process.env.JSPM_BIN) {
       console.error(err);
       process.exit(1);
     });
-    return;
   }
+  return;
 }
 
-const { default: cli } = require('../dist/cli');
+const { default: cli } = require('../lib/cli');
 process.once('unhandledRejection', err => {
   console.error(err);
   process.exit(1);
