+++ src/install/publish.ts
@@ -5,9 +5,8 @@ import { PackageName } from './package';
 import tar = require('tar-fs');
 import spdxLicenses = require('spdx-license-list/simple');
 import fs = require('graceful-fs');
-import path = require('path');
 import { Readable } from 'stream';
-import mm = require('micromatch');
+import { throwFieldMissing, throwFieldInvalid, createFilesFilter } from '../utils/pjson';
 
 interface PublishOptions {
   otp?: string;
@@ -15,78 +14,15 @@ interface PublishOptions {
   public?: boolean | void;
 }
 
-function throwFieldMissing (fieldName: string) {
-  throw new JspmUserError(`Package configuration must specify a ${bold(fieldName)} field to publish.`);
-}
-function throwFieldInvalid (value: string, expected: string) {
-  throw new JspmUserError(`${bold(JSON.stringify(value))} is not a valid ${expected}.`);
-}
-
-const alwaysIncludeFile = /^(readme|changelog|license|licence|notice)\.[^\.\/]+$/i;
-// (we ignore all dot files)
-const alwaysIgnoreFile = /^\.|^(CVS|npm-debug\.log|node_modules|jspm_packages|config\.gypi|[^\.]+\.orig|package-lock\.json|jspm\.json|jspm\.json\.lock|package\.json\.lock)$/;
-
-export function pack (project: Project, files: void | string[], ignore: void | string[]) {
-  const ignoresRes: RegExp[] = [];
-  const filesRes: RegExp[] = [];
-  const foldersRes: RegExp[] = [];
-  if (files) {
-    for (let pattern of files) {
-      pattern = pattern.replace(/\\/g, '/');
-      const parts = pattern.split('/');
-      while (parts.length > 1) {
-        parts.pop();
-        foldersRes.push(mm.makeRe(parts.join('/')));
-      }
-      filesRes.push(mm.makeRe(pattern));
-      filesRes.push(mm.makeRe(pattern + (pattern[pattern.length - 1] !== '/' ? '/' : '') + '**'));
-    }
-  }
-  if (ignore) {
-    for (let pattern of ignore) {
-      pattern = pattern.replace(/\\/g, '/');
-      ignoresRes.push(mm.makeRe(pattern));
-      ignoresRes.push(mm.makeRe(pattern + (pattern[pattern.length - 1] === '/' ? '/' : '') + '**'));
-    }
-  }
+export function pack (project: Project, ignoreFilter: (file: string) => boolean) {
   const includedFiles: string[] = [];
   const tarStream = <Readable>tar.pack(project.projectPath, <any>{
     fs,
     ignore (file: string) {
-      const relFile = path.relative(project.projectPath, file).replace(/\\/g, '/');
-
-      // always included
-      if (relFile.indexOf('/') === -1) {
-        if (relFile === 'package.json' || alwaysIncludeFile.test(relFile)) {
-          includedFiles.push(relFile);
-          return false;
-        }
-      }
-      // always ignored
-      if (alwaysIgnoreFile.test(path.basename(relFile)))
+      if (ignoreFilter(file))
         return true;
-
-      // "files"
-      if (files) {
-        if (filesRes.some(re => re.test(relFile))) {
-          if (ignoresRes.some(re => re.test(relFile)))
-            return true;
-          includedFiles.push(relFile);
-          return false;
-        }
-        // folder of "files" pattern
-        if (foldersRes.some(re => re.test(relFile))) {
-          includedFiles.push(relFile);
-          return false;
-        }
-        return true;
-      }
-      else {
-        if (ignoresRes.some(re => re.test(relFile)))
-          return true;
-        includedFiles.push(relFile);
-        return false;
-      }
+      includedFiles.push(file);
+      return false;
     },
     map (header) {
       header.name = 'package' + (header.name === '.' ? '' : '/' + header.name);
@@ -132,45 +68,7 @@ export default async function publish (project: Project, opts: PublishOptions) {
   delete pjson.devDependencies;
   delete pjson.overrides;
 
-  // We support package.json "ignore", falling back to .npmignore, .gitignore
-  let ignore = pjson.ignore;
-  let ignoreIsFile = false;
-  if (pjson.ignore instanceof Array === false) {
-    if ('ignore' in pjson)
-      throwFieldInvalid('ignore', 'array of patterns');
-    try {
-      const npmignore = await new Promise((resolve, reject) => 
-        fs.readFile(path.join(project.projectPath, '.npmignore'), (err, source) => err ? reject(err) : resolve(source))
-      );
-      ignoreIsFile = true;
-      ignore = npmignore.toString().trim().split('\n').map(item => item.trim()).filter(item => item[0] !== '#');
-    }
-    catch (e) {
-      if (e.code !== 'ENOENT')
-        throw e;
-      try {
-        const gitignore = await new Promise((resolve, reject) =>
-          fs.readFile(path.join(project.projectPath, '.gitignore'), (err, source) => err ? reject(err) : resolve(source))
-        );
-        ignoreIsFile = true;
-        ignore = gitignore.toString().trim().split('\n').map(item => item.trim()).filter(item => item[0] !== '#');
-      }
-      catch (e) {
-        if (e.code !== 'ENOENT')
-          throw e;
-      }
-    }
-  }
-
-  if ('files' in pjson && pjson.files instanceof Array === false) {
-    throwFieldInvalid('files', 'array of patterns');
-  }
-
-  // we dont ignore if there are files entries and the ignore came from a file (npm rule)
-  if (ignoreIsFile && pjson.files)
-    ignore = [];
-
-  const { files, stream } = pack(project, pjson.files, ignore);
+  const { files, stream } = pack(project, await createFilesFilter(project.projectPath, pjson.files, pjson.ignore));
 
   // log the files
   let logged = false;
